requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile
#
-e file:.
    # via -r requirements.in
aiohttp==3.8.0
    # via kopf
aiosignal==1.2.0
    # via aiohttp
asn1crypto==1.4.0
    # via
    #   certbuilder
    #   oscrypto
async-timeout==4.0.1
    # via aiohttp
attrs==21.2.0
    # via
    #   aiohttp
    #   pytest
build==0.7.0
    # via -r requirements.in
cachetools==4.2.4
    # via google-auth
certbuilder==0.14.2
    # via estoperator
certifi==2021.10.8
    # via
    #   kubernetes
    #   requests
cffi==1.15.0
    # via cryptography
charset-normalizer==2.0.7
    # via
    #   aiohttp
    #   requests
click==8.0.3
    # via kopf
cryptography==36.0.0
    # via estoperator
frozenlist==1.2.0
    # via
    #   aiohttp
    #   aiosignal
google-auth==2.3.3
    # via kubernetes
idna==3.3
    # via
    #   requests
    #   yarl
iniconfig==1.1.1
    # via pytest
iso8601==1.0.0
    # via kopf
kopf==1.35.2
    # via estoperator
kubernetes==19.15.0
    # via estoperator
multidict==5.2.0
    # via
    #   aiohttp
    #   yarl
oauthlib==3.1.1
    # via requests-oauthlib
oscrypto==1.2.1
    # via certbuilder
packaging==21.2
    # via
    #   build
    #   pytest
pep517==0.12.0
    # via build
pluggy==1.0.0
    # via pytest
py==1.11.0
    # via pytest
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
pycparser==2.21
    # via cffi
pyparsing==2.4.7
    # via packaging
pytest==6.2.5
    # via -r requirements.in
python-dateutil==2.8.2
    # via kubernetes
python-json-logger==2.0.2
    # via kopf
pyyaml==6.0
    # via
    #   kopf
    #   kubernetes
requests==2.26.0
    # via
    #   kubernetes
    #   requests-oauthlib
requests-oauthlib==1.3.0
    # via kubernetes
rsa==4.7.2
    # via google-auth
six==1.16.0
    # via
    #   google-auth
    #   kubernetes
    #   python-dateutil
toml==0.10.2
    # via pytest
tomli==1.2.2
    # via
    #   build
    #   pep517
typing-extensions==3.10.0.2
    # via
    #   async-timeout
    #   kopf
urllib3==1.26.7
    # via
    #   kubernetes
    #   requests
websocket-client==1.2.1
    # via kubernetes
wheel==0.37.0
    # via -r requirements.in
yarl==1.7.2
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# setuptools
